Fix updating remote repositories.
Updating remote repositories was not working
at all.

This is caused by JFrog's API failing when doing
an "update" and depending on the package type,
it will fail with an unhelpful "null" error
message.

Specifically it appears to be tied to it not likely specific "options"
that are only relevant to particular types of remote repos.

Below is a list of 7 flags,
that will cause a "generic" remoteRepo update API call to fail.

* vcsType (vcs only)
* vcsGitProvider (vcs only)
* vcsGitDownloadUrl (vcs only)
* feedContextPath (nuget only)
* downloadContextPath (nuget only)
* v3FeedUrl (nuget only)
* bowerRegistryUrl (bower only)

So the solution is for update calls for remote repo, to only include these flags if we are dealing with those particular types of repo: vcs/nuget/bower .

fix atlassian/terraform-provider-artifactory#52